pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mlcbase"
version = "1.1.0.rc1"
authors = [
    {name = "Weiming Chen", email = "wm_chen@yeah.net"},
    {name = "Yuanshuang Sun", email = "sunsun5544@126.com"},
    {name = "Zilin Yang", email = "876497115@qq.com"}
]
description = "The base module of all MuLingCloud modules and applications."
readme = {"file" = "README.md", content-type = "text/markdown"}
requires-python = ">=3.7"
dependencies = [
    "colorama",
    "loguru",
    "pytz",
    "prettytable",
    "pyyaml",
    "xmltodict",
    "rsa",
    "pycryptodome",
    "paramiko",
    "PyMySQL",
    "hvac"
]
license = { file = "LICENSE" }
classifiers = [
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    
    "Programming Language :: Python"
]

[project.urls]
Homepage = "https://www.mulingcloud.com"
Repository = "https://github.com/wmchen/mlcbase.git"